refactor: moved hardcoded values into constants file
